fix: resolve remaining issues
