+++ lib/desktop_screens_new/transfer_screen/file_transfer_screen.dart
@@ -1,6 +1,5 @@
 import 'dart:async';
 import 'dart:io';
-import 'dart:typed_data';
 
 import 'package:at_contact/at_contact.dart';
 import 'package:at_contacts_flutter/at_contacts_flutter.dart';
@@ -10,11 +9,9 @@ import 'package:atsign_atmosphere_pro/data_models/file_transfer.dart';
 import 'package:atsign_atmosphere_pro/data_models/file_transfer_status.dart';
 import 'package:atsign_atmosphere_pro/dekstop_services/desktop_image_picker.dart';
 import 'package:atsign_atmosphere_pro/desktop_screens_new/common_widgets/file_tile.dart';
-import 'package:atsign_atmosphere_pro/desktop_screens_new/contacts_screen/widgets/add_contacts_screen.dart';
-import 'package:atsign_atmosphere_pro/desktop_screens_new/groups_screen/widgets/desktop_add_group.dart';
 import 'package:atsign_atmosphere_pro/desktop_screens_new/notification/notification_icon.dart';
+import 'package:atsign_atmosphere_pro/desktop_screens_new/transfer_screen/widgets/add_atsigns_widget.dart';
 import 'package:atsign_atmosphere_pro/desktop_screens_new/transfer_screen/widgets/add_contact_tile.dart';
-import 'package:atsign_atmosphere_pro/screens/common_widgets/common_button.dart';
 import 'package:atsign_atmosphere_pro/screens/history/widgets/file_recipients.dart';
 import 'package:atsign_atmosphere_pro/services/common_utility_functions.dart';
 import 'package:atsign_atmosphere_pro/services/snackbar_service.dart';
@@ -66,14 +63,6 @@ class _FileTransferScreenState extends State<FileTransferScreen> {
   }
 
   List<GroupContactsModel> selectedContacts = [];
-
-  void addSelectedContact(GroupContactsModel contact) {
-    if (isSelected(contact)) {
-      selectedContacts.removeWhere((element) => element == contact);
-    } else {
-      selectedContacts.add(contact);
-    }
-  }
 
   sendFileWithFileBin(List<GroupContactsModel> contactList) async {
     bool isFilesReady = true;
@@ -666,402 +655,17 @@ class _FileTransferScreenState extends State<FileTransferScreen> {
       barrierColor: Colors.transparent,
       barrierDismissible: true,
       builder: (context) {
-        return StatefulBuilder(builder: (context, setDialogState) {
-          if (searchController.text.isEmpty) {
-            filteredContactList = [...GroupService().allContacts];
-            setDialogState(() {});
-          }
-
-          return Dialog(
-            insetPadding: EdgeInsets.zero,
-            alignment: Alignment.centerRight,
-            elevation: 5.0,
-            clipBehavior: Clip.hardEdge,
-            child: Container(
-              padding: EdgeInsets.symmetric(horizontal: 30),
-              decoration: BoxDecoration(
-                color: Theme.of(context).scaffoldBackgroundColor,
-                borderRadius: BorderRadius.circular(20),
-              ),
-              clipBehavior: Clip.hardEdge,
-              width: 400.toWidth,
-              child: Stack(
-                children: [
-                  SingleChildScrollView(
-                    child: Column(
-                      crossAxisAlignment: CrossAxisAlignment.start,
-                      children: [
-                        SizedBox(height: 20),
-                        Row(
-                          children: [
-                            InkWell(
-                              onTap: () {
-                                if (Navigator.canPop(context)) {
-                                  Navigator.pop(context);
-                                }
-                              },
-                              child: Icon(
-                                Icons.arrow_back_ios,
-                                size: 25,
-                              ),
-                            ),
-                            SizedBox(width: 15),
-                            Text(
-                              "Add atSigns",
-                              style: TextStyle(
-                                fontSize: 20,
-                                fontWeight: FontWeight.bold,
-                              ),
-                            ),
-                          ],
-                        ),
-                        SizedBox(height: 30),
-                        Padding(
-                          padding: const EdgeInsets.symmetric(horizontal: 16),
-                          child: Row(
-                            mainAxisAlignment: MainAxisAlignment.spaceBetween,
-                            children: [
-                              CommonButton(
-                                'Add contact',
-                                () async {
-                                  await showAddContactDialog()
-                                      .then((value) async {
-                                    if (value) {
-                                      if (searchController.text.isNotEmpty) {
-                                        searchController.clear();
-                                      }
-                                      await GroupService()
-                                          .fetchGroupsAndContacts(
-                                              isDesktop: true);
-                                      filteredContactList = [
-                                        ...GroupService().allContacts
-                                      ];
-                                      setDialogState(() {});
-                                    }
-                                  });
-                                },
-                                color: Color(0xFFF07C50),
-                                border: 20,
-                                height: 40,
-                                width: 136,
-                                fontSize: 18,
-                                removePadding: true,
-                              ),
-                              CommonButton(
-                                'Add group',
-                                () async {
-                                  await showAddGroupDialog()
-                                      .then((value) async {
-                                    if (value) {
-                                      if (searchController.text.isNotEmpty) {
-                                        searchController.clear();
-                                      }
-                                      await GroupService()
-                                          .fetchGroupsAndContacts(
-                                              isDesktop: true);
-                                      filteredContactList = [
-                                        ...GroupService().allContacts
-                                      ];
-                                      setDialogState(() {});
-                                    }
-                                  });
-                                },
-                                color: Color(0xFFF07C50),
-                                border: 20,
-                                height: 40,
-                                width: 136,
-                                fontSize: 18,
-                                removePadding: true,
-                              ),
-                            ],
-                          ),
-                        ),
-                        SizedBox(height: 30),
-                        Row(
-                          children: [
-                            Expanded(
-                              child: Container(
-                                decoration: BoxDecoration(
-                                  color: Colors.white,
-                                  borderRadius: BorderRadius.circular(5),
-                                ),
-                                child: TextField(
-                                  onChanged: (value) {
-                                    filteredContactList = [];
-                                    for (var contact
-                                        in GroupService().allContacts) {
-                                      if (contact?.contactType ==
-                                          ContactsType.CONTACT) {
-                                        if ((contact?.contact?.atSign
-                                                ?.contains(value)) ??
-                                            false) {
-                                          filteredContactList.add(contact);
-                                        }
-                                      } else if (contact?.contactType ==
-                                          ContactsType.GROUP) {
-                                        if ((contact?.group?.groupName
-                                                ?.contains(value)) ??
-                                            false) {
-                                          filteredContactList.add(contact);
-                                        }
-                                      }
-                                    }
-                                    setDialogState(() {});
-                                  },
-                                  controller: searchController,
-                                  decoration: InputDecoration(
-                                    isDense: true,
-                                    border: InputBorder.none,
-                                    contentPadding: EdgeInsets.symmetric(
-                                        horizontal: 20, vertical: 20),
-                                    hintText: "Search...",
-                                    suffixIcon: Icon(
-                                      Icons.search,
-                                      size: 25,
-                                    ),
-                                  ),
-                                ),
-                              ),
-                            ),
-                            SizedBox(width: 10),
-                            // InkWell(
-                            //   onTap: () {},
-                            //   child: CircleAvatar(
-                            //     backgroundColor: ColorConstants.MILD_GREY,
-                            //     child: Icon(
-                            //       Icons.verified_outlined,
-                            //       size: 25,
-                            //     ),
-                            //   ),
-                            // ),
-                          ],
-                        ),
-
-                        SizedBox(height: 30),
-
-                        // contact list
-                        ListView.builder(
-                          shrinkWrap: true,
-                          itemCount: filteredContactList.length,
-                          physics: NeverScrollableScrollPhysics(),
-                          itemBuilder: (context, index) {
-                            var groupContactModel = filteredContactList[index];
-                            late AtContact contact;
-                            var isTrusted = false;
-                            Uint8List? byteImage;
-
-                            if (groupContactModel?.contactType ==
-                                ContactsType.CONTACT) {
-                              contact = groupContactModel!.contact!;
-                              for (var ts in trustedContacts) {
-                                if (ts.atSign == (contact.atSign)) {
-                                  isTrusted = true;
-                                }
-                              }
-                              if (initialLetter != contact.atSign?[1]) {
-                                initialLetter = contact.atSign?[1] ?? "";
-                              } else {
-                                initialLetter = "";
-                              }
-                              byteImage = CommonUtilityFunctions()
-                                  .getCachedContactImage(
-                                contact.atSign!,
-                              );
-                            } else {
-                              if ((groupContactModel
-                                          ?.group?.groupName?.isNotEmpty ??
-                                      false) &&
-                                  initialLetter !=
-                                      groupContactModel?.group?.groupName?[0]) {
-                                initialLetter =
-                                    groupContactModel?.group?.groupName?[0] ??
-                                        "";
-                              } else {
-                                initialLetter = "";
-                              }
-                            }
-
-                            return Column(
-                              children: [
-                                initialLetter != ""
-                                    ? Row(
-                                        children: [
-                                          Text(
-                                            initialLetter,
-                                            style: TextStyle(
-                                              fontWeight: FontWeight.bold,
-                                              fontSize: 20,
-                                              color: Color(0xFF717171),
-                                            ),
-                                          ),
-                                          SizedBox(
-                                            width: 20,
-                                          ),
-                                          Expanded(
-                                            child: Divider(
-                                              thickness: 1,
-                                            ),
-                                          ),
-                                        ],
-                                      )
-                                    : SizedBox(),
-                                Padding(
-                                  padding: const EdgeInsets.only(
-                                      left: 15, top: 5, bottom: 5),
-                                  child: InkWell(
-                                    onTap: () {
-                                      addSelectedContact(groupContactModel!);
-                                      setDialogState(() {});
-                                    },
-                                    child: groupContactModel?.contactType ==
-                                            ContactsType.CONTACT
-                                        ? AddContactTile(
-                                            title: contact.atSign,
-                                            subTitle: contact.tags?["nickname"],
-                                            image: byteImage,
-                                            showImage: byteImage != null,
-                                            isSelected:
-                                                isSelected(groupContactModel!),
-                                            showDivider: true,
-                                            isTrusted: isTrusted,
-                                          )
-                                        : AddContactTile(
-                                            title: groupContactModel
-                                                ?.group?.groupName,
-                                            subTitle:
-                                                '${groupContactModel?.group?.members?.length} member(s)',
-                                            image: byteImage,
-                                            showImage: byteImage != null,
-                                            isSelected:
-                                                isSelected(groupContactModel!),
-                                            showDivider: true,
-                                            isTrusted: false,
-                                          ),
-                                  ),
-                                ),
-                              ],
-                            );
-                          },
-                        ),
-                        SizedBox(height: 100),
-                      ],
-                    ),
-                  ),
-                  Align(
-                    alignment: Alignment.bottomCenter,
-                    child: InkWell(
-                      onTap: () {
-                        context.read<FileTransferProvider>().selectedContacts =
-                            selectedContacts;
-                        context.read<FileTransferProvider>().notify();
-                        if (Navigator.canPop(context)) {
-                          Navigator.pop(context);
-                        }
-                      },
-                      child: Container(
-                        decoration: BoxDecoration(
-                          color: Colors.black,
-                          borderRadius: BorderRadius.circular(5),
-                        ),
-                        padding:
-                            EdgeInsets.symmetric(horizontal: 20, vertical: 20),
-                        margin: EdgeInsets.only(bottom: 20),
-                        width: double.maxFinite,
-                        child: Text(
-                          "Add atSigns ${selectedContacts.length}",
-                          style: TextStyle(
-                            color: Colors.white,
-                            fontSize: 18,
-                          ),
-                          textAlign: TextAlign.center,
-                        ),
-                      ),
-                    ),
-                  ),
-                ],
-              ),
-            ),
-          );
-        });
+        return AddAtSignsWidget(
+          trustedContacts: trustedContacts,
+          selectedContacts: selectedContacts,
+          addSelectedContactList: (value) {
+            selectedContacts = value;
+            context.read<FileTransferProvider>().selectedContacts =
+                selectedContacts;
+            context.read<FileTransferProvider>().notify();
+          },
+        );
       },
     );
-  }
-
-  Future<bool> showAddContactDialog() async {
-    bool shouldRefresh = false;
-    await showDialog(
-        context: context,
-        barrierColor: Colors.transparent,
-        barrierDismissible: true,
-        builder: (context) {
-          return Dialog(
-            insetPadding: EdgeInsets.zero,
-            alignment: Alignment.centerRight,
-            elevation: 5.0,
-            clipBehavior: Clip.hardEdge,
-            child: Container(
-              decoration: BoxDecoration(
-                color: Theme.of(context).scaffoldBackgroundColor,
-                borderRadius: BorderRadius.circular(20),
-              ),
-              clipBehavior: Clip.hardEdge,
-              width: 400.toWidth,
-              child: DesktopAddContactScreen(
-                onBack: (value) async {
-                  Navigator.pop(context);
-                  if (value) {
-                    shouldRefresh = value;
-                  }
-                },
-              ),
-            ),
-          );
-        });
-    return shouldRefresh;
-  }
-
-  Future<bool> showAddGroupDialog() async {
-    bool shouldRefresh = false;
-    await showDialog(
-        context: context,
-        barrierColor: Colors.transparent,
-        barrierDismissible: true,
-        builder: (context) {
-          return Dialog(
-            insetPadding: EdgeInsets.zero,
-            alignment: Alignment.centerRight,
-            elevation: 5.0,
-            clipBehavior: Clip.hardEdge,
-            child: Container(
-              decoration: BoxDecoration(
-                color: Theme.of(context).scaffoldBackgroundColor,
-                borderRadius: BorderRadius.circular(20),
-              ),
-              clipBehavior: Clip.hardEdge,
-              width: 400.toWidth,
-              child: DesktopAddGroup(
-                onDoneTap: (value) {
-                  Navigator.pop(context);
-                  if (value) {
-                    shouldRefresh = value;
-                  }
-                },
-              ),
-            ),
-          );
-        });
-    return shouldRefresh;
-  }
-
-  bool isSelected(GroupContactsModel groupContactsModel) {
-    for (GroupContactsModel contact in selectedContacts) {
-      if ((groupContactsModel.contactType == ContactsType.CONTACT &&
-              contact.contact == groupContactsModel.contact) ||
-          (groupContactsModel.contactType == ContactsType.GROUP &&
-              contact.group == groupContactsModel.group)) {
-        return true;
-      }
-    }
-    return false;
   }
 }
